+++ dear-imgui/src/widget/tooltip.rs
@@ -121,6 +121,7 @@ impl Ui {
     /// Returns true if the last item is being hovered by mouse with specific flags.
     #[doc(alias = "IsItemHovered")]
     pub fn is_item_hovered_with_flags(&self, flags: crate::HoveredFlags) -> bool {
+        crate::utils::validate_item_hovered_flags("Ui::is_item_hovered_with_flags()", flags);
         unsafe { sys::igIsItemHovered(flags.bits()) }
     }